Moving form and each card into their own component
